requirements.txt
JPype1-py3
